+++ tests/test_preprocessings/test_preprocessings.py
@@ -3,12 +3,14 @@ from label_binarizer import label_binarizer
 from label_encoder import label_encoder
 from standard_scaler import standard_scaler
 from binarizer import binarizer
+from function_transformer import function_transformer
 
 PREPROCESSINGS = [one_hot_encoder,
                   label_binarizer,
                   label_encoder,
                   standard_scaler,
-                  binarizer]
+                  binarizer,
+                  function_transformer]
 
 def test_full():
     for pre in PREPROCESSINGS: