+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- `FunctionTransformer` preprocessing model
 - `Binarizer` preprocessing model
 ### Changed
 - preprocessing params fulfilled in `pymilo_param`